requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
markovify==0.9.4
mecab-python3==1.0.5
oauthlib==3.2.0
plac==1.3.5
pycodestyle==2.8.0
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
toml==0.10.2
tqdm==4.64.0
tweepy==4.8.0
Unidecode==1.3.4
unidic==1.1.0
urllib3==1.26.9
wasabi==0.9.1
